Checkpoint: implemented general List parseCSVFile(path) method - CSV file fix.
Given a path to a CSV file, it serialize rows and fields comma separated
into a List<String[]> which is the returned type. This method can be
largely used to load other game components in the inizializeGame method.
The caller method must surround a parseCSVFile call with try/catch on a
IOException.